Various bugfixes.
Now the list updates correctly after the databse has been changed.
Only those items are counted in the nutrition totals which have the isAdded property set to 1.
On deletion of items, the forms page in popped off the stack and the application is moved to the home screen.
Checkboxes work correctly now.